Add confirm delete bookings

+++ frontend/src/components/UserBookings/DeleteBookingButton.js
@@ -5,13 +5,15 @@ import * as bookingsActions from '../../store/bookings';
 export default function DeleteBookingButton({ bookingId }) {
     const dispatch = useDispatch();
     const handleDeleteClick = async () => {
-        try {
-            await dispatch(bookingsActions.deleteBooking(bookingId))
-        }
+        if (window.confirm(`Please verify you would like to delete your booking by clicking "OK". This action cannot be undone.`)) {
+            try {
+                await dispatch(bookingsActions.deleteBooking(bookingId))
+            }
 
-        catch (res) {
-            const data = await res.json();
-            if (data.message) return alert(data.message);
+            catch (res) {
+                const data = await res.json();
+                if (data.message) return alert(data.message);
+            }
         }
     }
